DEV: fix after merging develop into branch re: updated gas limit and px uint type

+++ contracts/upgrade-mock/EnigmaV2.sol
@@ -315,8 +315,8 @@ contract EnigmaV2 is EnigmaStorage, EnigmaEvents, Getters, Ownable {
     */
     function createDeploymentTaskRecord(
         bytes32 _inputsHash,
-        uint _gasLimit,
-        uint _gasPx,
+        uint64 _gasLimit,
+        uint64 _gasPx,
         uint _firstBlockNumber,
         uint _nonce
     )
@@ -338,8 +338,8 @@ contract EnigmaV2 is EnigmaStorage, EnigmaEvents, Getters, Ownable {
     */
     function createTaskRecord(
         bytes32 _inputsHash,
-        uint _gasLimit,
-        uint _gasPx,
+        uint64 _gasLimit,
+        uint64 _gasPx,
         uint _firstBlockNumber
     )
     public
@@ -360,8 +360,8 @@ contract EnigmaV2 is EnigmaStorage, EnigmaEvents, Getters, Ownable {
     */
     function createTaskRecords(
         bytes32[] memory _inputsHashes,
-        uint[] memory _gasLimits,
-        uint[] memory _gasPxs,
+        uint64[] memory _gasLimits,
+        uint64[] memory _gasPxs,
         uint _firstBlockNumber
     )
     public
